Add Widget Bases.
