+++ tripal/src/Plugin/Field/TripalWidgetBase.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Drupal\tripal\Plugin\Field;
+
+use Drupal\Core\Field\WidgetBase;
+
+/**
+ * A Tripal-based entity field widget.
+ */
+abstract class TripalWidgetBase extends WidgetBase {
+
+}
